services/order-service/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
pymysql==1.1.1
cryptography==42.0.8
werkzeug==3.0.3
pyjwt==2.8.0
